🚚 Move subscribe logic to usePaywallModal
